feat(list-reorder): make changes from feedback
